refactor extract functions to utils.ts
